+++ model/Projectile.java
@@ -1,12 +1,11 @@
 public class Projectile {
     private int damage;
     private double speed;
-    private int row;
+    private double row;
     private double col;
 
     public void move() {
     }
-
     public void hitTarget(Zombie target) {
     }
 }
